requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyrogram>=2.0.0
tgcrypto>=1.2.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
torf>=4.0.0
